+++ apps/web/src/app/layout.tsx
@@ -30,7 +30,7 @@ export default function RootLayout({
   return (
     <html lang="en" className="dark">
       <body
-        className={`${inter.className} bg-background text-base text-foreground`}
+        className={`${inter.className} bg-background text-foreground text-base`}
       >
         <AppProvider>{children}</AppProvider>
       </body>